Tambah Fitur Cart

+++ src/main/java/com/prodemy/miniproject/dto/ProductDto.java
@@ -12,7 +12,7 @@ public class ProductDto {
     private Long id;
     private String productName;
     private Long categoryId;
-    private Long productPrice;
+    private double productPrice;
     private String productDescription;
     private String imageName;
